third_party/blink/web_tests/platform/win7/external/wpt/paint-timing/fcp-only/fcp-document-opacity-image-expected.txt

This is a testharness.js-based test.
FAIL Test that FCP after opacity change is not a larger value than LCP assert_less_than_equal: LCP is not smaller than FCP expected a number less than or equal to 23.7 but got 23.700000001117587
Harness: the test ran to completion.
